Add message.js module
